Remove type selection in VisualEditor dialog
Types were 'tag', 'switch', and 'variable'. As direct
insertion of wikitext does not work, all inserted code
from InsertMagic will be wrapped in <nowiki> tags. The only
way around this is to use VE commands, which will successively
be provided with the new tag factors and declared in the
mwvecommand attribute. For switches and variables, we do not have
this mechanism yet. Even if we get the respective commands,
switches will be located in the page options dialoge, and
variables can already be inserted in the templates dialog.

Please note this change only affects the MediaWiki VisualEditor
integration. The old dialoge for BlueSpice VisualEditor is still
fully functional.

Change-Id: Ia8bf6653542ee00450da030d4bc428784a7b6bcf